+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 
 ### v1.0.2 (unreleased)
 * Six Shooter now works with Sniper's Snipe properly (fixed in RoRR Modding Toolkit).
+* Ballistic Vest : Is now achievement locked
 * Overloaded Capacitor
     * First stack damage bonus: 20% -> 15%
     * Barrier out-of-combat delay time: 3s -> 5s
     * Barrier out-of-combat gain: 6% -> 8%
+* Stiletto : Is now achievement locked
 * Whimsical Star : Is now achievement locked
 * Added Relic Guard
 * Added Super Shield (temp name - also functionality changed from doc)